@@ -5,10 +5,10 @@ import 'package:plantify/common/colors.dart';
 import 'package:plantify/features/auth/presentation/bloc/toggle_password_visibility_cubit.dart';
 
 class TogglePasswordVisibility extends StatelessWidget {
-  final Function onClick;
+  final Function ()? onClick;
   final bool visible;
 
-  const TogglePasswordVisibility({Key key, this.visible, this.onClick})
+  const TogglePasswordVisibility({Key? key, required this.visible, this.onClick})
       : super(key: key);
 
   @override
