@@ -1,7 +1,7 @@
 mixin ValidationMixin {
-  bool isFieldEmpty(String fieldValue) => fieldValue?.isEmpty ?? true;
+  bool isFieldEmpty(String? fieldValue) => fieldValue?.isEmpty ?? true;
 
-  bool validateEmailAddress(String email) {
+  bool validateEmailAddress(String? email) {
     if (email == null) {
       return false;
     }
@@ -16,7 +16,7 @@ mixin ValidationMixin {
   }
 
 
-  bool validatePassword(String value) {
+  bool validatePassword(String? value) {
     if (value == null) return false;
 
     // (?=.*[A-Z])       // should contain at least one upper case
@@ -31,22 +31,21 @@ mixin ValidationMixin {
     return size && contains;
   }
 
-  bool validateCardNumber(String value) {
+  bool validateCardNumber(String? value) {
     if (value == null) return false;
     return value.length > 4;
   }
 
-  bool validField(String value) {
+  bool validField(String? value) {
     if (value == null) return false;
 
     return value.length > 3;
   }
 
-  bool validFullName(String value) {
+  bool validFullName(String? value) {
     if (value == null) return false;
 
-    Pattern namePattern = r"^[a-zA-Z]+(([',. -][a-zA-Z ])?[a-zA-Z]*)*$";
-    RegExp nameRegex = RegExp(namePattern, caseSensitive: false);
+    RegExp nameRegex = RegExp(r"^[a-zA-Z]+(([',. -][a-zA-Z ])?[a-zA-Z]*)*$", caseSensitive: false);
 
     return value.length > 2 && nameRegex.hasMatch(value);
   }
