@@ -3,14 +3,28 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:form_field_validator/form_field_validator.dart';
+import 'package:plantify/common/make_toast.dart';
 import 'package:plantify/common/validation/password_validator.dart';
 import 'package:plantify/common/widgets/basic_button.dart';
+import 'package:plantify/common/widgets/loading_widget.dart';
+import 'package:plantify/features/auth/presentation/bloc/register_screen_cubit.dart';
+import 'package:plantify/features/auth/presentation/bloc/result.dart';
 import 'package:plantify/features/auth/presentation/bloc/text_input_cubit.dart';
 import 'package:plantify/features/auth/presentation/bloc/toggle_password_visibility_cubit.dart';
 import 'package:plantify/features/auth/presentation/widgets/password_input.dart';
 import 'package:plantify/features/auth/presentation/widgets/text_input.dart';
 
 class RegisterScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (_) => RegisterScreenCubit(),
+      child: Register(),
+    );
+  }
+}
+
+class Register extends StatelessWidget {
   final textValidator = MultiValidator(
     [
       RequiredValidator(errorText: 'This field is required.'),
@@ -76,73 +90,106 @@ class RegisterScreen extends StatelessWidget {
               ],
             ),
           ),
-          Container(
-            width: MediaQuery.of(context).size.width,
-            padding: EdgeInsets.all(30.w),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                Form(
+          BlocBuilder<RegisterScreenCubit, Result>(
+            builder: (context, state) {
+              var cubit = context.watch<RegisterScreenCubit>();
+              return Container(
+                width: MediaQuery.of(context).size.width,
+                height: MediaQuery.of(context).size.height,
+                padding: EdgeInsets.all(30.w),
+                child: SingleChildScrollView(
+                  physics: BouncingScrollPhysics(),
+                  child: Container(
+                    height: MediaQuery.of(context).size.height - 2*30.w,
                     child: Column(
-                  children: [
-                    BlocProvider(
-                      create: (_) => TextInputCubit(textValidator),
-                      child: TextInput(
-                        hintText: 'First name',
-                      ),
-                    ),
-                    SizedBox(
-                      height: 20.h,
-                    ),
-                    BlocProvider(
-                      create: (_) => TextInputCubit(textValidator),
-                      child: TextInput(
-                        hintText: 'Last name',
-                      ),
-                    ),
-                    SizedBox(
-                      height: 20.h,
-                    ),
-                    BlocProvider(
-                      create: (_) => TextInputCubit(emailValidator),
-                      child: TextInput(
-                        hintText: 'Email',
-                        keyboardType: TextInputType.emailAddress,
-                      ),
-                    ),
-                    SizedBox(
-                      height: 20.h,
-                    ),
-                    MultiBlocProvider(
-                      providers: [
-                        BlocProvider(
-                          create: (_) => TextInputCubit(passwordValidator),
-                        ),
-                        BlocProvider(
-                          create: (_) => TogglePasswordVisibilityCubit(),
-                        ),
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      mainAxisSize: MainAxisSize.max,
+                      children: [
+                        Form(
+                            child: Column(
+                          children: [
+                            BlocProvider(
+                              create: (_) => TextInputCubit(textValidator),
+                              child: TextInput(
+                                hintText: 'First name',
+                              ),
+                            ),
+                            SizedBox(
+                              height: 20.h,
+                            ),
+                            BlocProvider(
+                              create: (_) => TextInputCubit(textValidator),
+                              child: TextInput(
+                                hintText: 'Last name',
+                              ),
+                            ),
+                            SizedBox(
+                              height: 20.h,
+                            ),
+                            BlocProvider(
+                              create: (_) => TextInputCubit(emailValidator),
+                              child: TextInput(
+                                hintText: 'Email',
+                                keyboardType: TextInputType.emailAddress,
+                              ),
+                            ),
+                            SizedBox(
+                              height: 20.h,
+                            ),
+                            MultiBlocProvider(
+                              providers: [
+                                BlocProvider(
+                                  create: (_) => TextInputCubit(passwordValidator),
+                                ),
+                                BlocProvider(
+                                  create: (_) => TogglePasswordVisibilityCubit(),
+                                ),
+                              ],
+                              child: PasswordInput(
+                                textInputAction: TextInputAction.done,
+                              ),
+                            ),
+                            SizedBox(
+                              height: 40.h,
+                            ),
+                            signUpButton(state, cubit) ?? Text(""),
+                            SizedBox(
+                              height: 40.h,
+                            ),
+                          ],
+                        ))
                       ],
-                      child: PasswordInput(
-                        textInputAction: TextInputAction.done,
-                      ),
                     ),
-                    SizedBox(
-                      height: 30.h,
-                    ),
-                    BasicButton(
-                      "Sign up",
-                      width: MediaQuery.of(context).size.width,
-                    ),
-                    SizedBox(
-                      height: 50.h,
-                    ),
-                  ],
-                ))
-              ],
-            ),
+                  ),
+                ),
+              );
+            },
           )
         ],
       ),
+    );
+  }
+
+  Widget button(RegisterScreenCubit cubit) => BasicButton(
+        "Sign up",
+        width: 250.w,
+        onClick: cubit.init,
+      );
+
+  Widget? signUpButton(Result state, RegisterScreenCubit cubit) {
+    return state.when(
+      () => button(cubit),
+      loading: () => SizedBox(
+        height: 50.w,
+        child: LoadingWidget(
+          size: 30.w,
+        ),
+      ),
+      error: (message) {
+        makeToast(message ?? "Something went wrong");
+        return button(cubit);
+      },
+      success: (data) => button(cubit),
     );
   }
 }
